Fixed Coulomb diagonal matrix elements bug.
The expression for the e3 matrix in Racah IV is actually for e3+Omega,
Omega is a diagonal matrix, so we need to subtract this from the m.e.
There was a check (i==j && e3!=0) -> (i==j) as need to subtract omega
from all elements, not just where e3+Omega!=0.

Also, for states of maximum multiplicity, the m.e. == -3omega. And the
tables of Racah do not cover this possibility. This was not considered
in the recurrence relations used for f6/f8, so some matrix elements of
f6 and f8 were wrong.